feat: 그룹 상세 화면과 장소 상세 화면 Navigation 연결

+++ client/presentation/src/main/java/com/andone/memorip/navigation/MemoripNavigator.kt
@@ -37,7 +37,7 @@ class MemoripNavigator(
         backStack.add(tab.route)
     }
 
-    fun navigateToPlaceDetail(placeId: Long) = backStack.navigateToPlaceDetail(placeId)
+    fun navigateToPlaceDetail(placeId: Int) = backStack.navigateToPlaceDetail(placeId)
 
     fun navigateToSelectCategory() = backStack.navigateToSelectCategory()
