+++ client/presentation/src/main/java/com/andone/memorip/navigation/Route.kt
@@ -22,4 +22,4 @@ data object SelectCategory : NavKey
 data class GroupDetail(val groupId: String) : NavKey
 
 @Serializable
-data class PlaceDetail(val placeId: Long = 0) : NavKey
+data class PlaceDetail(val placeId: Int = 0) : NavKey
