+++ client/presentation/src/main/java/com/andone/memorip/presentation/placedetail/PlaceDetailNavGraph.kt
@@ -6,7 +6,7 @@ import androidx.navigation3.runtime.NavBackStack
 import androidx.navigation3.runtime.NavKey
 import com.andone.memorip.navigation.PlaceDetail
 
-fun NavBackStack<NavKey>.navigateToPlaceDetail(placeId: Long) {
+fun NavBackStack<NavKey>.navigateToPlaceDetail(placeId: Int) {
     add(PlaceDetail(placeId))
 }
